requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
requests==2.28.2
aiogram-calendar==0.1.1
